Created first version of instrumentation tests
