+++ project/app/src/androidTest/java/com/nexte/nexte/FeedInstrumentalTest/CommentsFeedTest.java
@@ -1,4 +1,5 @@
-package com.nexte.nexte.PlayersListScene;
+package com.nexte.nexte.FeedInstrumentalTest;
+
 
 import android.support.test.espresso.ViewInteraction;
 import android.support.test.espresso.matcher.ViewMatchers;
@@ -21,31 +22,36 @@ import org.junit.runner.RunWith;
 
 import static android.support.test.espresso.Espresso.onView;
 import static android.support.test.espresso.action.ViewActions.click;
-import static android.support.test.espresso.action.ViewActions.scrollTo;
 import static android.support.test.espresso.matcher.ViewMatchers.isDisplayed;
-import static android.support.test.espresso.matcher.ViewMatchers.withClassName;
-import static android.support.test.espresso.matcher.ViewMatchers.withContentDescription;
 import static android.support.test.espresso.matcher.ViewMatchers.withId;
 import static android.support.test.espresso.matcher.ViewMatchers.withText;
 import static org.hamcrest.Matchers.allOf;
-import static org.hamcrest.Matchers.is;
 
 @LargeTest
 @RunWith(AndroidJUnit4.class)
-public class SendChallengeInstrumentationTest {
+public class CommentsFeedTest {
 
     @Rule
     public ActivityTestRule<MainActivity> mActivityTestRule = new ActivityTestRule<>(MainActivity.class);
 
     @Test
-    public void sendChallengeInstrumentationTest() {
+    public void commentsFeedTest() {
+        // Added a sleep statement to match the app's execution delay.
+        // The recommended way to handle such scenarios is to use Espresso idling resources:
+        // https://google.github.io/android-testing-support-library/docs/espresso/idling-resource/index.html
+        try {
+            Thread.sleep(700);
+        } catch (InterruptedException e) {
+            e.printStackTrace();
+        }
+
         ViewInteraction appCompatButton = onView(
-                allOf(ViewMatchers.withId(R.id.challangeButton), withText("Desafios"),
+                allOf(ViewMatchers.withId(R.id.navigationLogin), withText("Navegar pelo NEXTE!"),
                         childAtPosition(
                                 childAtPosition(
                                         withId(android.R.id.content),
                                         0),
-                                5),
+                                3),
                         isDisplayed()));
         appCompatButton.perform(click());
 
@@ -58,34 +64,16 @@ public class SendChallengeInstrumentationTest {
             e.printStackTrace();
         }
 
-        ViewInteraction appCompatImageView = onView(
-                allOf(withId(R.id.userPicture), withContentDescription("User picture"),
+        ViewInteraction appCompatTextView = onView(
+                allOf(withId(R.id.comments), withText("comentários"),
                         childAtPosition(
-                                childAtPosition(
-                                        withId(R.id.challengeRecyclerView),
-                                        1),
-                                0),
+                                allOf(withId(R.id.feedActivity),
+                                        childAtPosition(
+                                                withId(R.id.feedRecyclerView),
+                                                0)),
+                                12),
                         isDisplayed()));
-        appCompatImageView.perform(click());
-
-        ViewInteraction appCompatButton2 = onView(
-                allOf(withId(R.id.sendChallengeButton), withText("Enviar Desafio"),
-                        childAtPosition(
-                                childAtPosition(
-                                        withId(android.R.id.content),
-                                        0),
-                                5),
-                        isDisplayed()));
-        appCompatButton2.perform(click());
-
-        ViewInteraction appCompatButton3 = onView(
-                allOf(withId(android.R.id.button1), withText("Ok"),
-                        childAtPosition(
-                                childAtPosition(
-                                        withClassName(is("android.widget.ScrollView")),
-                                        0),
-                                3)));
-        appCompatButton3.perform(scrollTo(), click());
+        appCompatTextView.perform(click());
 
     }
