--- project/app/src/androidTest/java/com/nexte/nexte/MatchScene/MatchModelTestTest.kt
@@ -1,5 +1,0 @@
-package com.nexte.nexte.MatchScene
-
-import org.junit.Assert.*
-
-class MatchModelTestTest
