--- project/app/src/androidTest/java/com/nexte/nexte/PlayersListScene/ChallengeSentTest.java
@@ -1,142 +1,0 @@
-package com.nexte.nexte.PlayersListScene;
-
-
-import android.support.test.espresso.ViewInteraction;
-import android.support.test.espresso.matcher.ViewMatchers;
-import android.support.test.rule.ActivityTestRule;
-import android.support.test.runner.AndroidJUnit4;
-import android.view.View;
-import android.view.ViewGroup;
-import android.view.ViewParent;
-
-import com.nexte.nexte.MainActivity;
-import com.nexte.nexte.Player;
-import com.nexte.nexte.R;
-import com.nexte.nexte.UserSingleton;
-
-import org.hamcrest.Description;
-import org.hamcrest.Matcher;
-import org.hamcrest.TypeSafeMatcher;
-import org.hamcrest.core.IsInstanceOf;
-import org.junit.Rule;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-
-import static android.support.test.espresso.Espresso.onView;
-import static android.support.test.espresso.action.ViewActions.click;
-import static android.support.test.espresso.action.ViewActions.scrollTo;
-import static android.support.test.espresso.assertion.ViewAssertions.matches;
-
-import static android.support.test.espresso.matcher.ViewMatchers.isDisplayed;
-import static android.support.test.espresso.matcher.ViewMatchers.isEnabled;
-import static android.support.test.espresso.matcher.ViewMatchers.withClassName;
-import static android.support.test.espresso.matcher.ViewMatchers.withContentDescription;
-import static android.support.test.espresso.matcher.ViewMatchers.withId;
-import static android.support.test.espresso.matcher.ViewMatchers.withText;
-import static org.hamcrest.Matchers.allOf;
-import static org.hamcrest.Matchers.is;
-
-
-@RunWith(AndroidJUnit4.class)
-public class ChallengeSentTest {
-
-    @Rule
-    public ActivityTestRule<MainActivity> mActivityRule =
-            new ActivityTestRule<>(MainActivity.class);
-
-    @Test
-    public void successHasPlayersToChallenge(){
-        Player userInfo =
-                new Player("gabriel", 10, "","", "", "", 0, "");
-        UserSingleton.INSTANCE.setUserInformations(userInfo);
-        sleep();
-        onView(withText("Desafios")).perform(click());
-        sleep();
-        ViewInteraction appCompatImageView = onView(
-                allOf(ViewMatchers.withId(R.id.userPicture), withContentDescription("User picture"),
-                        childAtPosition(
-                                childAtPosition(
-                                        withId(R.id.challengeRecyclerView),
-                                        1),
-                                0),
-                        isDisplayed()));
-        appCompatImageView.perform(click());
-        sleep();
-        onView(withId(R.id.sendChallengeButton)).perform(click());
-        sleep();
-        onView(withText("Ok")).check(matches(isEnabled()));
-    }
-
-    @Test
-    public void noPlayersToMatch(){
-        Player userInfo =
-                new Player("gabriel", 1, "","", "", "", 0, "");
-        UserSingleton.INSTANCE.setUserInformations(userInfo);
-        ViewInteraction appCompatButton2 = onView(
-                allOf(withId(R.id.challangeButton), withText("Desafios"),
-                        childAtPosition(
-                                childAtPosition(
-                                        withId(android.R.id.content),
-                                        0),
-                                5),
-                        isDisplayed()));
-        appCompatButton2.perform(click());
-
-        // Added a sleep statement to match the app's execution delay.
-        // The recommended way to handle such scenarios is to use Espresso idling resources:
-        // https://google.github.io/android-testing-support-library/docs/espresso/idling-resource/index.html
-        try {
-            Thread.sleep(700);
-        } catch (InterruptedException e) {
-            e.printStackTrace();
-        }
-
-        ViewInteraction textView = onView(
-                allOf(withId(android.R.id.message), withText("Sem jogadores disponíveis.\nTente novamente mais tarde."),
-                        childAtPosition(
-                                childAtPosition(
-                                        IsInstanceOf.<View>instanceOf(android.widget.ScrollView.class),
-                                        0),
-                                1),
-                        isDisplayed()));
-        textView.check(matches(isDisplayed()));
-
-        ViewInteraction appCompatButton3 = onView(
-                allOf(withId(android.R.id.button1), withText("ok"),
-                        childAtPosition(
-                                childAtPosition(
-                                        withClassName(is("android.widget.ScrollView")),
-                                        0),
-                                3)));
-        appCompatButton3.perform(scrollTo(), click());
-    }
-
-    private static Matcher<View> childAtPosition(
-            final Matcher<View> parentMatcher, final int position) {
-
-        return new TypeSafeMatcher<View>() {
-            @Override
-            public void describeTo(Description description) {
-                description.appendText("Child at position " + position + " in parent ");
-                parentMatcher.describeTo(description);
-            }
-
-            @Override
-            public boolean matchesSafely(View view) {
-                ViewParent parent = view.getParent();
-                return parent instanceof ViewGroup && parentMatcher.matches(parent)
-                        && view.equals(((ViewGroup) parent).getChildAt(position));
-            }
-        };
-    }
-
-    private void sleep(){
-        try {
-            Thread.sleep(700);
-        } catch (InterruptedException e) {
-            e.printStackTrace();
-        }
-    }
-
-
-}
